style(esl-utils): fix docs and restructure
