+++ src/modules/esl-utils/misc/object.ts
@@ -1,22 +1,6 @@
+export * from './object/utils';
 export * from './object/types';
 export * from './object/path';
 export * from './object/merge';
 export * from './object/copy';
 export * from './object/compare';
-
-/** @returns first defined param */
-export function defined<T>(...params: T[]): T | undefined {
-  for (const param of params) {
-    if (param !== undefined) return param;
-  }
-}
-
-/** Find the closest property descriptor */
-export function getPropertyDescriptor(o: any, prop: PropertyKey): PropertyDescriptor | undefined {
-  let proto = o;
-  while (proto) {
-    const desc = Object.getOwnPropertyDescriptor(proto, prop);
-    if (desc) return desc;
-    proto = Object.getPrototypeOf(proto);
-  }
-}
