+++ src/modules/esl-utils/misc/object/test/compare.test.ts
@@ -1,0 +1,39 @@
+import {deepCompare} from '../compare';
+
+describe('misc/object: deepCompare', () => {
+  test.each([
+    [null, null],
+    [NaN, NaN],
+    ['', ''],
+    [{}, {}],
+    [{a: 1}, {a: 1}],
+    [{a: null}, {a: null}],
+    [{a: []}, {a: []}],
+    [{c: NaN}, {c: NaN}],
+    [{a: 1, c: Infinity}, {a: 1, c: Infinity}],
+    [{a: {b: 1}}, {a: {b: 1}}],
+    [{a: {b: {c: 1}}}, {a: {b: {c: 1}}}],
+    [[], []],
+    [[1], [1]],
+    [[1, 2, 3], [1, 2, 3]],
+    [[{}, {}], [{}, {}]],
+    [[{a: 1}, {b: ''}], [{a: 1}, {b: ''}]]
+  ])('%p should be equal to %p', (a: any, b: any) => expect(deepCompare(a, b)).toBe(true));
+
+  test.each([
+    [undefined, null],
+    [1, 2],
+    ['a', 'b'],
+    [{a: 1}, {b: 2}],
+    [{a: null}, {a: {}}],
+    [{a: {}}, {a: null}],
+    [{b: 1}, {b: 2}],
+    [{b: undefined}, {b: null}],
+    [{a: {c: {b: 1}}}, {a: {b: {c: 1}}}],
+    [{a: {a: 1, b: 1}}, {a: {b: 1}}],
+    [{a: {b: 1}}, {a: {a: 1, b: 1}}],
+    [[], [1]],
+    [[1], [1, 2]],
+    [[2, 1], [1, 2]]
+  ])('%p should not be equal to %p', (a: any, b: any) => expect(deepCompare(a, b)).toBe(false));
+});
